Mark newly created directories on overlay as opaque.
This is an optimization that will prevent needless lookups on
lower layers to merge. Because if mkdir(2) does succeed, then
none of the lower layers have a file at that position, otherwise
mkdir(2) would have failed with EEXIST.

We observe a similar behavior on linux. See the following:
```
$ sudo mount -t overlay -o lowerdir=lower,upperdir=upper,workdir=workdir none overlay
$ mkdir -p overlay/dir/dir1
$ mkdir -p lower/dir/dir2
$ ls overlay/dir/
dir1
```

But not sure if Linux actually marks upper/dir as opaque.

PiperOrigin-RevId: 456416625